Move Data Loaders to data_loader module
