+++ athenet/__init__.py
@@ -1,4 +1,5 @@
 from network import *
+import data_loader
 import layers
 import utils
 import models
