+++ athenet/data_loader/__init__.py
@@ -1,0 +1,3 @@
+from data_loader import *
+from mnist_data_loader import *
+from imagenet_data_loader import *
