+++ athenet/utils/__init__.py
@@ -1,7 +1,4 @@
 from constants import *
 from misc import *
-from data_loader import *
-from mnist_data_loader import *
-from imagenet_data_loader import *
 from run_algorithm import *
 from results_maintenance import *
